leo_parser
Module parser

The parser to convert Leo code text into a Program AST type.

This module contains the parse() function which calls the underlying tokenize() method to create a new program AST.

Modules

context 🔒

expression 🔒

file 🔒

statement 🔒

type_ 🔒

Functions

parse

Creates a new program from a given file path and source code text.