Module leo_parser::tokenizer::lexer

Structs

SpannedToken

Functions

check_address

Returns true if the given string looks like Aleo address. This method DOES NOT check if the address is valid on-chain.

eat

Returns the length of the given wanted string if the string can be eaten, otherwise returns None. A string can be eaten if its bytes are at the front of the given input array.

eat_identifier

Returns a new StrTendril string if an identifier can be eaten, otherwise returns None. An identifier can be eaten if its bytes are at the front of the given input_tendril string.