Enum leo_parser::tokenizer::token::Token

pub enum Token {

    CommentLine(String),
    CommentBlock(String),
    StaticString(String),
    Identifier(Symbol),
    Integer(String),
    True,
    False,
    AddressLit(String),
    WhiteSpace,
    Not,
    And,
    AndAssign,
    Or,
    OrAssign,
    BitAnd,
    BitAndAssign,
    BitOr,
    BitOrAssign,
    Eq,
    NotEq,
    Lt,
    LtEq,
    Gt,
    GtEq,
    Add,
    AddAssign,
    Sub,
    SubAssign,
    Mul,
    MulAssign,
    Div,
    DivAssign,
    Pow,
    PowAssign,
    Rem,
    RemAssign,
    Assign,
    LeftParen,
    RightParen,
    LeftSquare,
    RightSquare,
    LeftCurly,
    RightCurly,
    Comma,
    Dot,
    DotDot,
    Semicolon,
    Colon,
    DoubleColon,
    Question,
    Arrow,
    BigArrow,
    Shl,
    ShlAssign,
    Shr,
    ShrAssign,
    Underscore,
    BitXor,
    BitXorAssign,
    At,
    Address,
    Bool,
    Field,
    Group,
    Scalar,
    String,
    I8,
    I16,
    I32,
    I64,
    I128,
    U8,
    U16,
    U32,
    U64,
    U128,
    Record,
    Circuit,
    Console,
    Const,
    Constant,
    Decrement,
    Else,
    Finalize,
    For,
    Function,
    If,
    Import,
    In,
    Increment,
    Let,
    Mapping,
    Public,
    Return,
    SelfLower,
    Static,
    Eof,

}

Represents all valid Leo syntax tokens.

Variants

CommentLine(String)

CommentBlock(String)

StaticString(String)

Identifier(Symbol)

Integer(String)

True

False

AddressLit(String)

WhiteSpace

Not

And

AndAssign

Or

OrAssign

BitAnd

BitAndAssign

BitOr

BitOrAssign

Eq

NotEq

Lt

LtEq

Gt

GtEq

Add

AddAssign

Sub

SubAssign

Mul

MulAssign

Div

DivAssign

Pow

PowAssign

Rem

RemAssign

Assign

LeftParen

RightParen

LeftSquare

RightSquare

LeftCurly

RightCurly

Comma

Dot

DotDot

Semicolon

Colon

DoubleColon

Question

Arrow

BigArrow

Shl

ShlAssign

Shr

ShrAssign

Underscore

BitXor

BitXorAssign

At

Address

Bool

Field

Group

Scalar

String

I8

I16

I32

I64

I128

U8

U16

U32

U64

U128

Record

Circuit

Console

Const

Constant

Decrement

Else

Finalize

For

Function

If

Import

In

Increment

Let

Mapping

Public

Return

SelfLower

Static

Eof

Implementations

impl Token

pub fn is_keyword(&self) -> bool

Returns true if the self token equals a Leo keyword.

pub fn keyword_to_symbol(&self) -> Option<Symbol>

Converts self to the corresponding Symbol if it is_keyword.

impl Token

fn eat_integer(
input: &mut Peekable<impl Iterator<Item = char>>
) -> Result<(usize, Token)>

Returns a tuple: [(integer length, integer token)] if an integer can be eaten, otherwise returns None. An integer can be eaten if its bytes are at the front of the given input string.

pub(crate) fn eat(input: &str) -> Result<(usize, Token)>

Returns a tuple: [(token length, token)] if the next token can be eaten, otherwise returns None. The next token can be eaten if the bytes at the front of the given input string can be scanned into a token.