Module leo_parser::parser

The parser to convert Leo code text into an Program AST type.

This module contains the parse() method which calls the underlying tokenize() method to create a new program ast.

Modules

context 🔒

expression 🔒

file 🔒

input 🔒

statement 🔒

type_ 🔒

Functions

parse

Creates a new program from a given file path and source code text.

parse_input

Parses an input file at the given file path and source code text.