Struct leo_parser::parser::context::ParserContext

pub(crate) struct ParserContext<'a> {
    pub(crate) handler: &'a Handler,
    tokens: Vec<SpannedToken>,
    pub(crate) token: SpannedToken,
    pub(crate) prev_token: SpannedToken,
    pub(crate) disallow_circuit_construction: bool,
}

Stores a program in tokenized format plus additional context. May be converted into a Program AST by parsing all tokens.

Fields
handler: &'a Handler

Handler used to side-channel emit errors from the parser.

tokens: Vec<SpannedToken>

All un-bumped tokens.

token: SpannedToken

The current token, i.e., if p.tokens = ['3', *, '4'], then after a p.bump(), we’ll have p.token = '3'.

prev_token: SpannedToken

The previous token, i.e., if p.tokens = ['3', *, '4'], then after two p.bump()s, we’ll have p.token = '*' and p.prev_token = '3'.

disallow_circuit_construction: bool

true if parsing an expression for if and loop statements – means circuit inits are not legal

Implementations

impl<'a> ParserContext<'a>

pub fn new(handler: &'a Handler, tokens: Vec<SpannedToken>) -> Self

Returns a new ParserContext type given a vector of tokens.

pub(crate) fn bump(&mut self)

Advances the parser cursor by one token.

So e.g., if we had previous = A, current = B, and tokens = [C, D, E], then after p.bump(), the state will be previous = B, current = C, and tokens = [D, E].

pub(super) fn check(&self, tok: &Token) -> bool

Checks whether the current token is tok.

pub(super) fn check_int(&self) -> bool

Checks whether the current token is a Token::Int(_).

pub(super) fn eat(&mut self, token: &Token) -> bool

Returns true if the next token is equal to the given token. Advances the parser to the next token.

pub(super) fn look_ahead<'s, R>(
&'s self,
dist: usize,
looker: impl FnOnce(&'s SpannedToken) -> R
) -> R

Look-ahead dist tokens of self.token and get access to that token there. When dist == 0 then the current token is looked at.

pub(super) fn emit_err(&self, err: ParserError)

Emit the error err.

pub(super) fn emit_warning(&self, warning: ParserWarning)

Emit the error err.

pub(crate) fn has_next(&self) -> bool

Returns true if the next token exists.

fn mk_ident_prev(&self, name: Symbol) -> Identifier

At the previous token, return and make an identifier with name.

pub(super) fn eat_identifier(&mut self) -> Option<Identifier>

Eats the next token if its an identifier and returns it.

pub(super) fn expect_identifier(&mut self) -> Result<Identifier>

Expects an Identifier, or errors.

pub fn eat_integer(&mut self) -> Result<(PositiveNumber, Span)>

Removes the next token if it is a [Token::Integer(_)] and returns it, or None if the next token is not a [Token::Integer(_)] or if the next token does not exist.

pub(super) fn eat_any(&mut self, tokens: &[Token]) -> bool

Eats any of the given tokens, returning true if anything was eaten.

fn unexpected<T>(&self, expected: impl Display) -> Result<T>

Returns an unexpected error at the current token.

pub(super) fn expect(&mut self, token: &Token) -> Result<Span>

Eats the expected token, or errors.

pub(super) fn expect_any(&mut self, tokens: &[Token]) -> Result<Span>

Eats one of the expected tokens, or errors.

pub(super) fn parse_list<T>(
&mut self,
delimiter: Delimiter,
sep: Option<Token>,
inner: impl FnMut(&mut Self) -> Result<Option<T>>
) -> Result<(Vec<T>, bool, Span)>

Parses a list of Ts using inner The opening and closing delimiters are bra and ket, and elements in the list are optionally separated by sep. When (list, true) is returned, sep was a terminator.

pub(super) fn parse_paren_comma_list<T>(
&mut self,
f: impl FnMut(&mut Self) -> Result<Option<T>>
) -> Result<(Vec<T>, bool, Span)>

Parse a list separated by , and delimited by parens.

pub(super) fn peek_is_left_par(&self) -> bool

Returns true if the current token is (.

impl ParserContext<'_>

pub(crate) fn parse_expression(&mut self) -> Result<Expression>

Returns an Expression AST node if the next token is an expression. Includes circuit init expressions.

pub(super) fn parse_conditional_expression(&mut self) -> Result<Expression>

Returns an Expression AST node if the next tokens represent a ternary expression. May or may not include circuit init expressions.

Otherwise, tries to parse the next token using [parse_boolean_or_expression].

fn bin_expr(
left: Expression,
right: Expression,
op: BinaryOperation
) -> Expression

Constructs a binary expression left op right.

fn parse_bin_expr(
&mut self,
tokens: &[Token],
f: impl FnMut(&mut Self) -> Result<Expression>
) -> Result<Expression>

Parses a left-associative binary expression <left> token <right> using f for left/right. The token is translated to op in the AST.

fn parse_boolean_or_expression(&mut self) -> Result<Expression>

Returns an Expression AST node if the next tokens represent a binary OR expression.

Otherwise, tries to parse the next token using [parse_boolean_and_expression].

fn parse_boolean_and_expression(&mut self) -> Result<Expression>

Returns an Expression AST node if the next tokens represent a binary AND expression.

Otherwise, tries to parse the next token using [parse_equality_expression].

fn eat_bin_op(&mut self, tokens: &[Token]) -> Option<BinaryOperation>

Eats one of binary operators matching any in tokens.

fn parse_ordering_expression(&mut self) -> Result<Expression>

Returns an Expression AST node if the next tokens represent a binary relational expression: less than, less than or equals, greater than, greater than or equals.

Otherwise, tries to parse the next token using [parse_additive_expression].

fn parse_equality_expression(&mut self) -> Result<Expression>

Returns an Expression AST node if the next tokens represent a binary equals or not equals expression.

Otherwise, tries to parse the next token using [parse_ordering_expression].

fn parse_bitwise_exclusive_or_expression(&mut self) -> Result<Expression>

Returns an Expression AST node if the next tokens represent a bitwise exclusive or expression.

Otherwise, tries to parse the next token using [parse_bitwise_inclusive_or_expression].

fn parse_bitwise_inclusive_or_expression(&mut self) -> Result<Expression>

Returns an Expression AST node if the next tokens represent a bitwise inclusive or expression.

Otherwise, tries to parse the next token using [parse_bitwise_and_expression].

fn parse_bitwise_and_expression(&mut self) -> Result<Expression>

Returns an Expression AST node if the next tokens represent a bitwise and expression.

Otherwise, tries to parse the next token using [parse_shift_expression].

fn parse_shift_expression(&mut self) -> Result<Expression>

Returns an Expression AST node if the next tokens represent a shift left or a shift right expression.

Otherwise, tries to parse the next token using [parse_additive_expression].

fn parse_additive_expression(&mut self) -> Result<Expression>

Returns an Expression AST node if the next tokens represent a binary addition or subtraction expression.

Otherwise, tries to parse the next token using [parse_mul_div_pow_expression].

fn parse_multiplicative_expression(&mut self) -> Result<Expression>

Returns an Expression AST node if the next tokens represent a binary multiplication, division, or modulus expression.

Otherwise, tries to parse the next token using [parse_exponential_expression].

fn parse_exponential_expression(&mut self) -> Result<Expression>

Returns an Expression AST node if the next tokens represent a binary exponentiation expression.

Otherwise, tries to parse the next token using [parse_unary_expression].

pub(super) fn parse_unary_expression(&mut self) -> Result<Expression>

Returns an Expression AST node if the next tokens represent a unary not, negate, or bitwise not expression.

Otherwise, tries to parse the next token using [parse_postfix_expression].

fn parse_method_call_expression(
&mut self,
receiver: Expression,
method: Identifier
) -> Result<Expression>

Returns an Expression AST node if the next tokens represent a method call expression.

fn parse_associated_access_expression(
&mut self,
circuit_name: Expression
) -> Result<Expression>

Returns an Expression AST node if the next tokens represent a static access expression.

pub(crate) fn parse_expr_tuple(
&mut self
) -> Result<(Vec<Expression>, bool, Span)>

Parses a tuple of Expression AST nodes.

fn parse_postfix_expression(&mut self) -> Result<Expression>

Returns an Expression AST node if the next tokens represent an array access, circuit member access, function call, or static function call expression.

Otherwise, tries to parse the next token using [parse_primary_expression].

fn parse_tuple_expression(&mut self) -> Result<Expression>

Returns an Expression AST node if the next tokens represent a tuple initialization expression or an affine group literal.

fn peek_group_coordinate(&self, dist: &mut usize) -> Option<GroupCoordinate>

Returns a reference to the next token if it is a GroupCoordinate, or None if the next token is not a GroupCoordinate.

fn eat_group_partial(&mut self) -> Option<Result<GroupTuple>>

Removes the next two tokens if they are a pair of GroupCoordinate and returns them, or None if the next token is not a GroupCoordinate.

fn parse_circuit_member(&mut self) -> Result<CircuitVariableInitializer>

pub fn parse_circuit_init_expression(
&mut self,
identifier: Identifier
) -> Result<Expression>

Returns an Expression AST node if the next tokens represent a circuit initialization expression. let foo = Foo { x: 1u8 };

fn parse_primary_expression(&mut self) -> Result<Expression>

Returns an Expression AST node if the next token is a primary expression:

Literals: field, group, unsigned integer, signed integer, boolean, address

Aggregate types: array, tuple

Identifiers: variables, keywords

self

Returns an expression error if the token cannot be matched.

impl ParserContext<'_>

pub fn parse_program(&mut self) -> Result<Program>

Returns a Program AST if all tokens can be consumed and represent a valid Leo program.

fn unexpected_item(token: &SpannedToken) -> ParserError

pub(super) fn parse_import(&mut self) -> Result<(Identifier, Program)>

Parses an import statement import foo.leo;.

fn parse_circuit_members(&mut self) -> Result<(Vec<CircuitMember>, Span)>

Returns a Vec<CircuitMember> AST node if the next tokens represent a circuit member variable or circuit member function or circuit member constant.

pub(super) fn parse_typed_ident(&mut self) -> Result<(Identifier, Type)>

Parses IDENT: TYPE.

fn parse_const_member_variable_declaration(&mut self) -> Result<CircuitMember>

Returns a CircuitMember AST node if the next tokens represent a circuit member static constant.

fn parse_member_variable_declaration(&mut self) -> Result<CircuitMember>

Returns a CircuitMember AST node if the next tokens represent a circuit member variable.

fn parse_member_function_declaration(&mut self) -> Result<CircuitMember>

Returns a CircuitMember AST node if the next tokens represent a circuit member function.

pub(super) fn parse_circuit(&mut self) -> Result<(Identifier, Circuit)>

Parses a circuit or record definition, e.g., circit Foo { ... } or record Foo { ... }.

pub(super) fn parse_function_parameter_mode(&mut self) -> Result<ParamMode>

Returns a ParamMode AST node if the next tokens represent a function parameter mode.

fn parse_function_parameter(&mut self) -> Result<FunctionInput>

Returns a FunctionInput AST node if the next tokens represent a function parameter.

fn peek_is_function(&self) -> bool

Returns true if the next token is Function or if it is a Const followed by Function. Returns false otherwise.

fn parse_function(&mut self) -> Result<(Identifier, Function)>

Returns an [(Identifier, Function)] AST node if the next tokens represent a function name and function definition.

impl ParserContext<'_>

pub(crate) fn parse_input(&mut self) -> Result<InputAst>

Returns a [ParsedInputFile] struct filled with the data acquired in the file.

fn parse_section(&mut self) -> Result<Section>

Parses particular section in the Input or State file. [<identifier>] <...definition> Returns Section.

fn parse_input_definition(&mut self) -> Result<Definition>

Parses a single parameter definition: <identifier> : <type> = <expression>; Returns Definition.

impl ParserContext<'_>

pub(crate) fn parse_statement(&mut self) -> Result<Statement>

Returns a Statement AST node if the next tokens represent a statement.

fn parse_assign_statement(&mut self) -> Result<Statement>

Returns a Block AST node if the next tokens represent a assign, or expression statement.

pub(super) fn parse_block(&mut self) -> Result<Block>

Returns a Block AST node if the next tokens represent a block of statements.

fn parse_return_statement(&mut self) -> Result<ReturnStatement>

Returns a ReturnStatement AST node if the next tokens represent a return statement.

fn parse_conditional_statement(&mut self) -> Result<ConditionalStatement>

Returns a ConditionalStatement AST node if the next tokens represent a conditional statement.

fn parse_loop_statement(&mut self) -> Result<IterationStatement>

Returns an IterationStatement AST node if the next tokens represent an iteration statement.

fn parse_console_args(&mut self) -> Result<ConsoleArgs>

Returns a ConsoleArgs AST node if the next tokens represent a formatted string.

fn parse_console_statement(&mut self) -> Result<ConsoleStatement>

Returns a ConsoleStatement AST node if the next tokens represent a console statement.

pub(super) fn parse_definition_statement(
&mut self
) -> Result<DefinitionStatement>

Returns a DefinitionStatement AST node if the next tokens represent a definition statement.

impl ParserContext<'_>

pub fn parse_primitive_type(&mut self) -> Result<(Type, Span)>

Returns a [(Type, Span)] tuple of AST nodes if the next token represents a primitive type. Also returns the span of the parsed token.

pub fn parse_type(&mut self) -> Result<(Type, Span)>

Returns a [(Type, Span)] tuple of AST nodes if the next token represents a type. Also returns the span of the parsed token.

fn parse_tuple_type(&mut self) -> Result<(Type, Span)>

Parses a type of form (ty_0, ty_1, ...).