Struct leo_parser::parser::context::ParserContext

pub struct ParserContext {
    tokens: Vec<SpannedToken>,
    end_span: Span,
    pub(crate) fuzzy_struct_state: bool,
}

Stores a program in tokenized format plus additional context. May be converted into a [Program] AST by parsing all tokens.

Fields
tokens: Vec<SpannedToken>end_span: Spanfuzzy_struct_state: bool
Implementations

impl ParserContext

pub fn new(tokens: Vec<SpannedToken>) -> Self

Returns a new ParserContext type given a vector of tokens.

pub fn eof(&self) -> LeoError

Returns an unexpected end of function [SyntaxError].

pub fn peek_next(&self) -> Result<&SpannedToken>

Returns a reference to the next next token or error if it does not exist.

pub fn peek(&self) -> Result<&SpannedToken>

Returns a reference to the next token or error if it does not exist.

pub fn peek_token(&self) -> Cow<'_, Token>

pub fn has_next(&self) -> bool

Returns true if the next token exists.

pub fn eat(&mut self, token: Token) -> Option<SpannedToken>

Removes the next token if it exists and returns it, or None if the next token does not exist.

pub fn backtrack(&mut self, token: SpannedToken)

Appends a token to the back of the vector.

pub fn eat_identifier(&mut self) -> Option<Identifier>

Removes the next token if it is a [Token::Ident(_)] and returns it, or None if the next token is not a [Token::Ident(_)] or if the next token does not exist.

fn peek_group_coordinate(&self, i: &mut usize) -> Option<GroupCoordinate>

Returns a reference to the next token if it is a [GroupCoordinate], or None if the next token is not a [GroupCoordinate].

pub fn eat_group_partial(
&mut self
) -> Option<Result<(GroupCoordinate, GroupCoordinate, Span)>>

Removes the next two tokens if they are a pair of [GroupCoordinate] and returns them, or None if the next token is not a [GroupCoordinate].

pub fn eat_int(&mut self) -> Option<(PositiveNumber, Span)>

Removes the next token if it is a [Token::Int(_)] and returns it, or None if the next token is not a [Token::Int(_)] or if the next token does not exist.

pub fn eat_any(&mut self, token: &[Token]) -> Option<SpannedToken>

Removes the next token if it exists and returns it, or None if the next token does not exist.

pub fn expect(&mut self, token: Token) -> Result<Span>

Returns the span of the next token if it is equal to the given Token, or error.

pub fn expect_oneof(&mut self, token: &[Token]) -> Result<SpannedToken>

Returns the span of the next token if it is equal to one of the given Tokens, or error.

pub fn expect_loose_identifier(&mut self) -> Result<Identifier>

Returns the [Identifier] of the next token if it is a keyword, [Token::Int(_)], or an [Identifier], or error.

pub fn expect_ident(&mut self) -> Result<Identifier>

Returns the [Identifier] of the next token if it is an [Identifier], or error.

pub fn expect_any(&mut self) -> Result<SpannedToken>

Returns the next token if it exists or return end of function.

impl ParserContext

pub fn parse_expression(&mut self) -> Result<Expression>

Returns an [Expression] AST node if the next token is an expression. Includes circuit init expressions.

pub fn parse_conditional_expression(&mut self) -> Result<Expression>

Returns an [Expression] AST node if the next tokens represent a ternary expression. May or may not include circuit init expressions.

Otherwise, tries to parse the next token using [parse_disjunctive_expression].

pub fn parse_disjunctive_expression(&mut self) -> Result<Expression>

Returns an [Expression] AST node if the next tokens represent a binary or expression.

Otherwise, tries to parse the next token using [parse_conjunctive_expression].

pub fn parse_conjunctive_expression(&mut self) -> Result<Expression>

Returns an [Expression] AST node if the next tokens represent a binary and expression.

Otherwise, tries to parse the next token using [parse_bit_or_expression].

pub fn parse_equality_expression(&mut self) -> Result<Expression>

Returns an [Expression] AST node if the next tokens represent a binary bitwise or expression.

Otherwise, tries to parse the next token using [parse_bit_xor_expression].

Returns an [Expression] AST node if the next tokens represent a binary bitwise xor expression.

Otherwise, tries to parse the next token using [parse_bit_and_expression].

Returns an [Expression] AST node if the next tokens represent a binary bitwise and expression.

Otherwise, tries to parse the next token using [parse_equality_expression].

Returns an [Expression] AST node if the next tokens represent a binary equals or not equals expression.

Otherwise, tries to parse the next token using [parse_ordering_expression].

pub fn parse_ordering_expression(&mut self) -> Result<Expression>

Returns an [Expression] AST node if the next tokens represent a binary relational expression: less than, less than or equals, greater than, greater than or equals.

Otherwise, tries to parse the next token using [parse_shift_expression].

pub fn parse_additive_expression(&mut self) -> Result<Expression>

Returns an [Expression] AST node if the next tokens represent a binary shift expression.

Otherwise, tries to parse the next token using [parse_additive_expression].

Returns an [Expression] AST node if the next tokens represent a binary addition or subtraction expression.

Otherwise, tries to parse the next token using [parse_mul_div_pow_expression].

pub fn parse_multiplicative_expression(&mut self) -> Result<Expression>

Returns an [Expression] AST node if the next tokens represent a binary multiplication, division, or modulus expression.

Otherwise, tries to parse the next token using [parse_exponential_expression].

pub fn parse_exponential_expression(&mut self) -> Result<Expression>

Returns an [Expression] AST node if the next tokens represent a binary exponentiation expression.

Otherwise, tries to parse the next token using [parse_cast_expression].

pub fn parse_cast_expression(&mut self) -> Result<Expression>

Returns an [Expression] AST node if the next tokens represent a type cast expression.

Otherwise, tries to parse the next token using [parse_unary_expression].

pub fn parse_unary_expression(&mut self) -> Result<Expression>

Returns an [Expression] AST node if the next tokens represent a unary not, negate, or bitwise not expression.

Otherwise, tries to parse the next token using [parse_postfix_expression].

pub fn parse_postfix_expression(&mut self) -> Result<Expression>

Returns an [Expression] AST node if the next tokens represent an array access, circuit member access, function call, or static function call expression.

Otherwise, tries to parse the next token using [parse_primary_expression].

pub fn parse_spread_or_expression(&mut self) -> Result<SpreadOrExpression>

Returns a [SpreadOrExpression] AST node if the next tokens represent an spread or expression.

This method should only be called in the context of an array access expression.

pub fn parse_circuit_expression(
&mut self,
identifier: Identifier
) -> Result<Expression>

Returns an [Expression] AST node if the next tokens represent an circuit initialization expression.

pub fn parse_tuple_expression(&mut self, span: &Span) -> Result<Expression>

Returns an [Expression] AST node if the next tokens represent an tuple initialization expression.

pub fn parse_array_expression(&mut self, span: &Span) -> Result<Expression>

Returns an [Expression] AST node if the next tokens represent an array initialization expression.

pub fn parse_primary_expression(&mut self) -> Result<Expression>

Returns an [Expression] AST node if the next token is a primary expression:

Literals: field, group, unsigned integer, signed integer, boolean, address

Aggregate types: array, tuple

Identifiers: variables, keywords

self

Returns an expression error if the token cannot be matched.

impl ParserContext

pub fn parse_program(&mut self) -> Result<Program>

Returns a [Program] AST if all tokens can be consumed and represent a valid Leo program.

pub fn parse_annotation(&mut self) -> Result<Annotation>

Returns an [Annotation] AST node if the next tokens represent a supported annotation.

pub fn parse_package_accesses(
&mut self,
span: &Span
) -> Result<Vec<PackageAccess>>

Returns a vector of [PackageAccess] AST nodes if the next tokens represent package access expressions within an import statement.

pub fn parse_package_access(&mut self) -> Result<PackageAccess>

Returns a [PackageAccess] AST node if the next tokens represent a package access expression within an import statement.

pub fn parse_package_name(&mut self) -> Result<Identifier>

Returns an [Identifier] AST node if the next tokens represent a valid package name.

pub fn parse_package_path(&mut self) -> Result<PackageOrPackages>

Returns a [PackageOrPackages] AST node if the next tokens represent a valid package import with accesses.

pub fn parse_import_statement(&mut self) -> Result<ImportStatement>

Returns a [ImportStatement] AST node if the next tokens represent an import statement.

pub fn parse_circuit_declaration(&mut self) -> Result<Vec<CircuitMember>>

Returns a [CircuitMember] AST node if the next tokens represent a circuit member variable or circuit member function.

pub fn parse_member_variable_declaration(
&mut self
) -> Result<(CircuitMember, bool)>

Returns a [CircuitMember] AST node if the next tokens represent a circuit member variable.

pub fn parse_member_function_declaration(&mut self) -> Result<CircuitMember>

Returns a [CircuitMember] AST node if the next tokens represent a circuit member function.

pub fn parse_circuit(&mut self) -> Result<(Identifier, Circuit)>

Returns an [(Identifier, Circuit)] tuple of AST nodes if the next tokens represent a circuit name and definition statement.

pub fn parse_function_parameters(&mut self) -> Result<FunctionInput>

Returns a [FunctionInput] AST node if the next tokens represent a function parameter.

pub fn parse_function_declaration(&mut self) -> Result<(Identifier, Function)>

Returns an [(Identifier, Function)] AST node if the next tokens represent a function name and function definition.

pub fn parse_global_const_declaration(
&mut self
) -> Result<(Vec<Identifier>, DefinitionStatement)>

Returns an [(String, DefinitionStatement)] AST node if the next tokens represent a global const definition statement and assignment.

pub fn parse_type_alias(&mut self) -> Result<(Identifier, Alias)>

Returns an [(String, Alias)] AST node if the next tokens represent a global const definition statement and assignment.

impl ParserContext

pub fn construct_assignee_access(
expr: Expression,
accesses: &mut Vec<AssigneeAccess>
) -> Result<Identifier>

Returns an [Identifier] AST node if the given [Expression] AST node evaluates to an identifier access. The access is stored in the given accesses.

pub fn construct_assignee(expr: Expression) -> Result<Assignee>

Returns an [Assignee] AST node from the given [Expression] AST node with accesses.

pub fn parse_statement(&mut self) -> Result<Statement>

Returns a [Statement] AST node if the next tokens represent a statement.

pub fn parse_assign_statement(&mut self) -> Result<Statement>

Returns a [Block] AST node if the next tokens represent a assign, or expression statement.

pub fn parse_block(&mut self) -> Result<Block>

Returns a [Block] AST node if the next tokens represent a block of statements.

pub fn parse_return_statement(&mut self) -> Result<ReturnStatement>

Returns a [ReturnStatement] AST node if the next tokens represent a return statement.

pub fn parse_conditional_statement(&mut self) -> Result<ConditionalStatement>

Returns a [ConditionalStatement] AST node if the next tokens represent a conditional statement.

pub fn parse_loop_statement(&mut self) -> Result<IterationStatement>

Returns an [IterationStatement] AST node if the next tokens represent an iteration statement.

pub fn parse_console_args(&mut self) -> Result<ConsoleArgs>

Returns a [ConsoleArgs] AST node if the next tokens represent a formatted string.

pub fn parse_console_statement(&mut self) -> Result<ConsoleStatement>

Returns a [ConsoleStatement] AST node if the next tokens represent a console statement.

pub fn parse_variable_name(
&mut self,
span: &SpannedToken
) -> Result<VariableName>

Returns a [VariableName] AST node if the next tokens represent a variable name with valid keywords.

pub fn parse_definition_statement(&mut self) -> Result<DefinitionStatement>

Returns a [DefinitionStatement] AST node if the next tokens represent a definition statement.

impl ParserContext

pub fn token_to_int_type(token: Token) -> Option<IntegerType>

Returns a [IntegerType] AST node if the given token is a supported integer type, or None.

pub fn parse_array_dimensions(&mut self) -> Result<Option<ArrayDimensions>>

Returns an [ArrayDimensions] AST node if the next tokens represent dimensions for an array type.

pub fn parse_type(&mut self) -> Result<(Type, Span)>

Returns a [(Type, Span)] tuple of AST nodes if the next token represents a type. Also returns the span of the parsed token.