Enum leo_parser::tokenizer::token::Token

pub enum Token {

    CommentLine(StrTendril),
    CommentBlock(StrTendril),
    StringLit(Vec<Char>),
    Ident(StrTendril),
    Int(StrTendril),
    True,
    False,
    AddressLit(StrTendril),
    CharLit(Char),
    At,
    Not,
    And,
    Or,
    Eq,
    NotEq,
    Lt,
    LtEq,
    Gt,
    GtEq,
    Add,
    Minus,
    Mul,
    Div,
    Exp,
    Assign,
    AddEq,
    MinusEq,
    MulEq,
    DivEq,
    ExpEq,
    LeftParen,
    RightParen,
    LeftSquare,
    RightSquare,
    LeftCurly,
    RightCurly,
    Comma,
    Dot,
    DotDot,
    DotDotDot,
    Semicolon,
    Colon,
    DoubleColon,
    Question,
    Arrow,
    Underscore,
    U8,
    U16,
    U32,
    U64,
    U128,
    I8,
    I16,
    I32,
    I64,
    I128,
    Field,
    Group,
    Bool,
    Address,
    Char,
    BigSelf,
    Input,
    LittleSelf,
    Import,
    As,
    Circuit,
    Console,
    Const,
    Else,
    For,
    Function,
    If,
    In,
    Let,
    Mut,
    Return,
    Static,
    Type,
    LengthOf,
    Eof,

}

Represents all valid Leo syntax tokens.

Variants

CommentLine(StrTendril)

CommentBlock(StrTendril)

StringLit(Vec<Char>)

Ident(StrTendril)

Int(StrTendril)

True

False

AddressLit(StrTendril)

CharLit(Char)

At

Not

And

Or

Eq

NotEq

Lt

LtEq

Gt

GtEq

Add

Minus

Mul

Div

Exp

Assign

AddEq

MinusEq

MulEq

DivEq

ExpEq

LeftParen

RightParen

LeftSquare

RightSquare

LeftCurly

RightCurly

Comma

Dot

DotDot

DotDotDot

Semicolon

Colon

DoubleColon

Question

Arrow

Underscore

U8

U16

U32

U64

U128

I8

I16

I32

I64

I128

Field

Group

Bool

Address

Char

BigSelf

Input

LittleSelf

Import

As

Circuit

Console

Const

Else

For

Function

If

In

Let

Mut

Return

Static

Type

LengthOf

Eof

Implementations

impl Token

pub fn is_keyword(&self) -> bool

Returns true if the self token equals a Leo keyword.

impl Token

fn eat_char(
input_tendril: StrTendril,
escaped: bool,
hex: bool,
unicode: bool
) -> Option<Char>

Returns a char if a character can be eaten, otherwise returns None.

fn eat_integer(input_tendril: &StrTendril) -> (usize, Option<Token>)

Returns a tuple: [(integer length, integer token)] if an integer can be eaten, otherwise returns None. An integer can be eaten if its bytes are at the front of the given input_tendril string.

fn utf8_byte_count(byte: u8) -> u8

Returns the number of bytes in an emoji via a bit mask.

pub(crate) fn eat(input_tendril: StrTendril) -> (usize, Option<Token>)

Returns a tuple: [(token length, token)] if the next token can be eaten, otherwise returns None. The next token can be eaten if the bytes at the front of the given input_tendril string can be scanned into a token.